+++ package/R/brar_normal.R
@@ -2,8 +2,8 @@
 #'     effect estimates
 #'
 #' @description This function computes Bayes factors, posterior probabilities,
-#'     and response-adaptive randomization probabilities in the setting of data
-#'     in the form of approximately normal effect estimates.
+#'     and response-adaptive randomization probabilities for data summarized by
+#'     approximately normal effect estimates.
 #'
 #' @param estimate Vector of effect estimates (e.g., a vector of mean
 #'     differences or log odds/hazard/rate ratios). Each estimate quantifies the
@@ -11,9 +11,9 @@
 #' @param sigma Covariance matrix of the effect estimate vector. In case, there
 #'     is only one effect estimate, this is the squared standard error of the
 #'     effect estimate
-#' @param pm Mean vector of the normal prior assigned to the effect under the
+#' @param pm Mean vector of the normal prior assigned to the effects under the
 #'     alternative. Defaults to \code{rep(0, length(estimate))}
-#' @param psigma Covariance matrix of the normal prior assigned to the effect
+#' @param psigma Covariance matrix of the normal prior assigned to the effects
 #'     under the alternative. In case, there is only one effect estimate, this
 #'     is the prior variance
 #' @param pH0 Prior probability of the point null hypothesis (i.e., all
@@ -25,7 +25,7 @@
 #'     of the null hypothesis and prior probabilities of control/treatment
 #'     superiority), \code{"BF_ij"} (Bayes factor matrix), \code{"posterior"}
 #'     (posterior probability of the null hypothesis and posterior probabilities
-#'     of control/treatment superiority), and "prand" (response-adaptive
+#'     of control/treatment superiority), and \code{"prand"} (response-adaptive
 #'     randomization probabilities).
 #'
 #' @author Samuel Pawel
@@ -197,6 +197,7 @@ brar_normal <- function(estimate, sigma, pm = rep(0, length(estimate)), psigma,
         ## compute BFs
         margdens <- c(margdensHm, margdensH0, resHp[,3])
         bfmat <- outer(X = margdens, Y = margdens, FUN = `/`)
+        diag(bfmat) <- 1
         Hpnames <- paste0("H+", seq(1, K))
         colnames(bfmat) <- rownames(bfmat) <- c("H-", "H0", Hpnames)
 
@@ -206,8 +207,33 @@ brar_normal <- function(estimate, sigma, pm = rep(0, length(estimate)), psigma,
         prior <- c(pHm, pH0, pHp)
         names(prior) <- c("H-", "H0", Hpnames)
 
+        ## ## compute posterior hypothesis probabilitites
+        ## post <- prior*margdens/sum(prior*margdens)
         ## compute posterior hypothesis probabilitites
-        post <- prior*margdens/sum(prior*margdens)
+        if (pH0 == 1) {
+            ## if Pr(H0) = 1, posterior prob of H0 is always 1 regardless of data
+            post <- c(0, 1, rep(0, K))
+        } else if (sum(margdens == 0) > 1) {
+            ## more than one marginal log likelihood = -Inf => all post probs are NaN
+            post <- rep(NaN, K + 2)
+        } else {
+            odds <- outer(X = prior, Y = prior, FUN = `/`)
+            diag(odds) <- 1
+            post <- 1/sapply(X = seq_len(ncol(bfmat)), FUN = function(i) {
+                sum(bfmat[,i]*odds[,i])
+            })
+        }
+        names(post) <- names(prior)
+    }
+    ## if prior is 0, posterior is always 0 regardless of data
+    ## set posterior manually to avoid numerical NaN issues
+    post[prior == 0] <- 0
+    ## if posterior prob of any hypothesis is 1, all other probs are 0
+    ## set manually to avoid some numerical NaN issues
+    post1 <- post == 1
+    if (any(post1) & sum(post1, na.rm = TRUE) == 1) {
+        post[!post1] <- 0
+        post[is.nan(post)] <- 0
     }
 
     ## compute randomization probabilities