+++ package/NEWS.md
@@ -1,4 +1,7 @@
 # brar 0.1
 
-- implemented point null Bayesian response-adaptive randomization for binomial and normal outcomes with `brar_binomial` and `brar_normal`
+- CRAN release: <https://CRAN.R-project.org/package=brar>, <https://doi.org/10.32614/CRAN.package.brar>
+- implemented null hypothesis Bayesian response-adaptive randomization for
+  binomial and normal outcomes with `brar_binomial` and `brar_normal` as
+  described in <https://doi.org/10.48550/arXiv.2510.01734>
 - package development started